requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
loguru==0.7.2
playwright==1.47.0
httpx==0.27.2
